Finished correct guess validation
Project is still buggy. Still working on setting up rooms properly so that they can be joined by players and
hosts in any order, and the secret word can be submitted at any time without messing anything up. However
for now, I moved on to making sure guess validation is actually complete so that even if we are unable to
fully complete the room portion, the project will still be presentable when we ensure players and hosts join
in a specific order. As is right now, it must go as follows. HOST JOINS -> PLAYER JOINS -> HOST ESTABLISHES AND
SENDS SECRET WORD -> PLAYER MAY GUESS WORD AS MANY TIMES AS IT TAKES.