Presentable version
This version of the project is presentable. It is NOT FINISHED.
I still need to implement a queue to handle player/host connections
because the boolean flags that are currently being used are not
sufficient. We need FIFO handling. However for the purposes of presenting
this version should do fine.